chore(interactive.explorer): add init module
