+++ src/erlab/interactive/explorer/_base_explorer.py
@@ -1,18 +1,3 @@
-"""Quickly browse and load ARPES data files with a file manager-like interface.
-
-.. image:: ../images/explorer_light.png
-    :align: center
-    :alt: Data explorer window in light mode
-    :class: only-light
-
-.. only:: format_html
-
-    .. image:: ../images/explorer_dark.png
-        :align: center
-        :alt: Data explorer window in dark mode
-        :class: only-dark
-"""
-
 from __future__ import annotations
 
 import os
@@ -1191,40 +1176,3 @@ class _DataExplorer(QtWidgets.QMainWindow):
                         if file_paths[0].is_dir()
                         else file_paths[0].parent
                     )
-
-
-def data_explorer(
-    directory: str | os.PathLike | None = None,
-    loader_name: str | None = None,
-    *,
-    execute: bool | None = None,
-) -> None:
-    """Start the data explorer.
-
-    Data explorer is a tool to browse and load ARPES data files with a file manager-like
-    interface. Data attributes of supported files can be quickly inspected, and can be
-    loaded into ImageToolManager for further analysis.
-
-    The data explorer can be started from the command line as a standalone application
-    with the following command:
-
-    .. code-block:: bash
-
-        python -m erlab.interactive.explorer
-
-    Also, it can be opened from the GUI by selecting "File" -> "Data Explorer" in
-    ImageToolManager.
-
-    Parameters
-    ----------
-    directory
-        Initial directory to display in the explorer.
-    loader_name
-        Name of the loader to use to load the data. The loader must be registered in
-        :attr:`erlab.io.loaders`.
-    """
-    with erlab.interactive.utils.setup_qapp(execute):
-        win = _DataExplorer(root_path=directory, loader_name=loader_name)
-        win.show()
-        win.raise_()
-        win.activateWindow()
